Updated gamma stuff: We don't use gamma ramps (in fact we can, we have SDL_SetGammaRamp), therefore eliminated V_CheckGamma & friends from V_UpdatePalette for PLATFORM_UNIX. Also killed unused function WarpPalette.
git-svn-id: file:///svn/uhexen2/code/trunk@512 eaf649e6-6c13-0410-b194-84221022858e

+++ hexen2/view.h
@@ -3,7 +3,7 @@
 extern	cvar_t		v_gamma;
 
 extern	byte		gammatable[256];	// palette is sent through this
-extern	byte		ramps[3][256];
+extern	unsigned short	ramps[3][256];
 extern float v_blend[4];
 
 extern cvar_t lcd_x;
